refactor(generator): common function to process templates
