+++ generator/internal/language/generate.go
@@ -1,0 +1,75 @@
+// Copyright 2024 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package language
+
+import (
+	"errors"
+	"fmt"
+	"os"
+	"path/filepath"
+
+	"github.com/cbroglie/mustache"
+)
+
+type mustacheProvider struct {
+	impl    func(string) (string, error)
+	dirname string
+}
+
+func (p *mustacheProvider) Get(name string) (string, error) {
+	return p.impl(filepath.Join(p.dirname, name) + ".mustache")
+}
+
+// GenerateFromRoot generates a number of files always using the root template
+// data as the input to the mustache template.
+//
+// This works well for languages like Go and Rust, where a single file can
+// contain all the clients or messages generated by an API. We anticipate that
+// other functions will be needed for languages like Java or C++, where it is
+// conventional to have a single class per file.
+//
+// TODO(#653) - `root any` will become `model *api.API` once the intermediate templatedata is gone.
+func GenerateFromRoot(outdir string, root any, provider templateProvider, generatedFiles []GeneratedFile) error {
+	var errs []error
+	for _, gen := range generatedFiles {
+		templateContents, err := provider(gen.TemplatePath)
+		if err != nil {
+			errs = append(errs, err)
+			continue
+		}
+		if outdir == "" {
+			wd, _ := os.Getwd()
+			outdir = wd
+		}
+		destination := filepath.Join(outdir, gen.OutputPath)
+		os.MkdirAll(filepath.Dir(destination), 0777) // Ignore errors
+		nestedProvider := mustacheProvider{
+			impl:    provider,
+			dirname: filepath.Dir(gen.TemplatePath),
+		}
+		s, err := mustache.RenderPartials(templateContents, &nestedProvider, root)
+		if err != nil {
+			errs = append(errs, err)
+			continue
+		}
+		if err := os.WriteFile(destination, []byte(s), os.ModePerm); err != nil {
+			errs = append(errs, err)
+		}
+	}
+	if len(errs) > 0 {
+		return fmt.Errorf("errors generating output files: %w", errors.Join(errs...))
+	}
+	return nil
+}
